requirements.txt
qdrant-client==1.12.1
sentence-transformers==3.3.1
pypdf==5.1.0
python-dotenv==1.0.1
langchain-text-splitters==1.1.0
